Masks: Use a macro for consistent handle detection
The macro DT_MASKS_WIDTH_HANDLE is used to define the size of a node and handle.
It is used also for distance to the cursor for selection, so that all mask types behave
consistently.
The value is DPI aware, so it should be the same on all screens at all zoom level.